envActionsAttachedToReportings
