--- backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/entities/mission/ActionTargetTypeEnum.kt
@@ -1,7 +1,0 @@
-package fr.gouv.cacem.monitorenv.domain.entities.mission
-
-enum class ActionTargetTypeEnum {
-    VEHICLE,
-    COMPANY,
-    INDIVIDUAL,
-}
