--- backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/entities/mission/FormalNoticeEnum.kt
@@ -1,7 +1,0 @@
-package fr.gouv.cacem.monitorenv.domain.entities.mission
-
-enum class FormalNoticeEnum {
-    YES,
-    NO,
-    PENDING,
-}
