--- backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/entities/mission/InfractionTypeEnum.kt
@@ -1,7 +1,0 @@
-package fr.gouv.cacem.monitorenv.domain.entities.mission
-
-enum class InfractionTypeEnum {
-    WAITING,
-    WITH_REPORT,
-    WITHOUT_REPORT,
-}
