--- backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/entities/mission/ThemeEntity.kt
@@ -1,7 +1,0 @@
-package fr.gouv.cacem.monitorenv.domain.entities.mission
-
-data class ThemeEntity(
-    val theme: String? = null,
-    val subThemes: List<String>? = listOf(),
-    val protectedSpecies: List<String>? = listOf(),
-)
