--- backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/entities/mission/VesselSizeEnum.kt
@@ -1,10 +1,0 @@
-@file:Suppress("ktlint:standard:enum-entry-name-case")
-
-package fr.gouv.cacem.monitorenv.domain.entities.mission
-
-enum class VesselSizeEnum {
-    LESS_THAN_12m,
-    FROM_12_TO_24m,
-    FROM_24_TO_46m,
-    MORE_THAN_46m,
-}
