+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/entities/mission/envAction/envActionControl/ActionTargetTypeEnum.kt
@@ -1,0 +1,7 @@
+package fr.gouv.cacem.monitorenv.domain.entities.mission.envAction.envActionControl
+
+enum class ActionTargetTypeEnum {
+    VEHICLE,
+    COMPANY,
+    INDIVIDUAL,
+}
