+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/entities/mission/envAction/envActionControl/infraction/FormalNoticeEnum.kt
@@ -1,0 +1,7 @@
+package fr.gouv.cacem.monitorenv.domain.entities.mission.envAction.envActionControl.infraction
+
+enum class FormalNoticeEnum {
+    YES,
+    NO,
+    PENDING,
+}
