+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/entities/mission/envAction/envActionControl/infraction/InfractionTypeEnum.kt
@@ -1,0 +1,7 @@
+package fr.gouv.cacem.monitorenv.domain.entities.mission.envAction.envActionControl.infraction
+
+enum class InfractionTypeEnum {
+    WAITING,
+    WITH_REPORT,
+    WITHOUT_REPORT,
+}
