+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/entities/mission/envAction/envActionControl/infraction/VesselSizeEnum.kt
@@ -1,0 +1,8 @@
+package fr.gouv.cacem.monitorenv.domain.entities.mission.envAction.envActionControl.infraction
+
+enum class VesselSizeEnum {
+    LESS_THAN_12m,
+    FROM_12_TO_24m,
+    FROM_24_TO_46m,
+    MORE_THAN_46m,
+}
