--- backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/api/adapters/bff/outputs/MissionDataOutput.kt
@@ -1,67 +1,0 @@
-package fr.gouv.cacem.monitorenv.infrastructure.api.adapters.bff.outputs
-
-import fr.gouv.cacem.monitorenv.domain.entities.controlUnit.LegacyControlUnitEntity
-import fr.gouv.cacem.monitorenv.domain.entities.mission.EnvActionEntity
-import fr.gouv.cacem.monitorenv.domain.entities.mission.MissionSourceEnum
-import fr.gouv.cacem.monitorenv.domain.entities.mission.MissionTypeEnum
-import fr.gouv.cacem.monitorenv.domain.use_cases.missions.dtos.MissionDTO
-import org.locationtech.jts.geom.MultiPolygon
-import java.time.ZonedDateTime
-
-data class MissionDataOutput(
-    val id: Int,
-    val missionTypes: List<MissionTypeEnum>,
-    val controlUnits: List<LegacyControlUnitEntity>? = listOf(),
-    val openBy: String? = null,
-    val closedBy: String? = null,
-    val observationsCacem: String? = null,
-    val observationsCnsp: String? = null,
-    val facade: String? = null,
-    val geom: MultiPolygon? = null,
-    val startDateTimeUtc: ZonedDateTime,
-    val endDateTimeUtc: ZonedDateTime? = null,
-    val envActions: List<EnvActionEntity>? = null,
-    val missionSource: MissionSourceEnum,
-    val isClosed: Boolean,
-    val hasMissionOrder: Boolean,
-    val isUnderJdp: Boolean,
-    val attachedReportingIds: List<Int>? = listOf(),
-    val attachedReportings: List<MissionAttachedReportingDataOutput>? = listOf(),
-    val detachedReportingIds: List<Int>? = listOf(),
-    val detachedReportings: List<MissionDetachedReportingDataOutput>? = listOf(),
-) {
-    companion object {
-        fun fromMissionDTO(dto: MissionDTO): MissionDataOutput {
-            requireNotNull(dto.mission.id) {
-                "a mission must have an id"
-            }
-
-            return MissionDataOutput(
-                id = dto.mission.id,
-                missionTypes = dto.mission.missionTypes,
-                controlUnits = dto.mission.controlUnits,
-                openBy = dto.mission.openBy,
-                closedBy = dto.mission.closedBy,
-                observationsCacem = dto.mission.observationsCacem,
-                observationsCnsp = dto.mission.observationsCnsp,
-                facade = dto.mission.facade,
-                geom = dto.mission.geom,
-                startDateTimeUtc = dto.mission.startDateTimeUtc,
-                endDateTimeUtc = dto.mission.endDateTimeUtc,
-                envActions = dto.mission.envActions,
-                missionSource = dto.mission.missionSource,
-                isClosed = dto.mission.isClosed,
-                hasMissionOrder = dto.mission.hasMissionOrder,
-                isUnderJdp = dto.mission.isUnderJdp,
-                attachedReportingIds = dto.attachedReportingIds,
-                attachedReportings = dto.attachedReportings?.map {
-                    MissionAttachedReportingDataOutput.fromReportingDTO(it)
-                },
-                detachedReportingIds = dto.detachedReportingIds,
-                detachedReportings = dto.detachedReportings?.map {
-                    MissionDetachedReportingDataOutput.fromReporting(it.reporting)
-                },
-            )
-        }
-    }
-}
