+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/api/adapters/bff/outputs/missions/MissionDataOutput.kt
@@ -1,0 +1,2 @@
+package fr.gouv.cacem.monitorenv.infrastructure.api.adapters.bff.outputs.missions
+
